+++ src/boxes/admin/admin_board/UserTotalStatistic.css
@@ -8,7 +8,7 @@
  float: left;
 }
 .userLocal-Total{
-  width: 60%;
+  width: 90%;
   margin-bottom: 20px;
 }
